Adding application beta version for Foogle App
Foogle is a platform where people can post about their home made food.
People can find other’s post and order food. Each post include name of
the dish, ingredients, location, image of the dish, deadline for
ordering, number of dishes left and price of the dish and a brief
description about the dish or any other message for the post. Newsfeed
page only shows summary of a post. When user selects a post detailed is
shown on following page with purchasing option. The searching of food
can be done based on location, ingredients, price, dish name or type
etc. Two types of posts will be there i.e. buying food and selling food.
The payment can be done through PayPal and credit card. All orders can
be found in Cart section of the application. These orders are the
confirmation of the orders one makes and can be used to identify the
buyer when they visit seller’s place to have food. A user must sign in
to our application with a valid email address and password. Each user
will have a profile where they can tell about their expertise in cooking
food. Profile also includes rating for each user from 0 to 5. The MVP of
our product is to be able to open an account, post about dish, buying
dish, getting order signature for buying or selling, an online payment
method (i.e., PayPal) which have already been implemented. We will be
working more on both user interface design and adding or removing few
features based on customer feedback on the beta version of our mobile
application. We plan to have web version of our app if we get positive
feedback on the beta version. Following are the screenshots of the beta
version of our mobile application.